requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
opencv-python-headless
numpy
pandas
pytest
httpx
python-dotenv

